% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_utils.R
\name{check_site_exists}
\alias{check_site_exists}
\title{Check if Site Exists}
\usage{
check_site_exists(con, site_id)
}
\arguments{
\item{con}{Database connection}

\item{site_id}{Site ID to check}
}
\value{
Logical TRUE if exists
}
\description{
Check if Site Exists
}
