% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_utils.R
\name{fetch_all_batch_scans}
\alias{fetch_all_batch_scans}
\title{Fetch All Batch Scans}
\usage{
fetch_all_batch_scans(con)
}
\arguments{
\item{con}{Database connection}
}
\value{
Data frame of all batch scans with site names
}
\description{
Fetch All Batch Scans
}
