% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_utils.R
\name{fetch_all_labels}
\alias{fetch_all_labels}
\title{Fetch All Labels}
\usage{
fetch_all_labels(con)
}
\arguments{
\item{con}{Database connection}
}
\value{
Data frame of all labels
}
\description{
Fetch All Labels
}
