% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_utils.R
\name{fetch_all_sites}
\alias{fetch_all_sites}
\title{Fetch All Sites}
\usage{
fetch_all_sites(con)
}
\arguments{
\item{con}{Database connection}
}
\value{
Data frame of all sites
}
\description{
Fetch All Sites
}
