% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_utils.R
\name{fetch_site_by_id}
\alias{fetch_site_by_id}
\title{Fetch Site by ID}
\usage{
fetch_site_by_id(con, site_id)
}
\arguments{
\item{con}{Database connection}

\item{site_id}{Site ID to fetch}
}
\value{
Data frame of matching site
}
\description{
Fetch Site by ID
}
