% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_utils.R
\name{generate_label_png}
\alias{generate_label_png}
\title{Generate Label PNG}
\usage{
generate_label_png(qr_text, filename)
}
\arguments{
\item{qr_text}{Text for QR code}

\item{filename}{Output filename}
}
\value{
Path to generated image
}
\description{
Generate Label PNG
}
