% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_utils.R
\name{get_next_site_id}
\alias{get_next_site_id}
\title{Get Next Sequential Site ID}
\usage{
get_next_site_id(con)
}
\arguments{
\item{con}{Database connection}
}
\value{
Next site ID (ST0001, ST0002, etc.)
}
\description{
Generates the next sequential site ID based on existing sites in database.
Starts with ST0001 if no sites exist.
}
