% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_utils.R
\name{initialize_database}
\alias{initialize_database}
\title{Initialize Database}
\usage{
initialize_database()
}
\value{
Path to the database file
}
\description{
Creates SQLite database with required schema for LIMS
}
