% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_utils.R
\name{insert_batch_scan}
\alias{insert_batch_scan}
\title{Insert Batch Scan}
\usage{
insert_batch_scan(con, site_id, scanner_input)
}
\arguments{
\item{con}{Database connection}

\item{site_id}{Site ID}

\item{scanner_input}{Scanned data}
}
\description{
Insert Batch Scan
}
