% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_utils.R
\name{insert_label}
\alias{insert_label}
\title{Insert Label}
\usage{
insert_label(con, label_id, site_id, sample_id, qr_code_path)
}
\arguments{
\item{con}{Database connection}

\item{label_id}{Label ID}

\item{site_id}{Site ID}

\item{sample_id}{Sample ID}

\item{qr_code_path}{Path to QR code image}
}
\description{
Insert Label
}
