% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_utils.R
\name{insert_site}
\alias{insert_site}
\title{Insert Site}
\usage{
insert_site(con, site_id, site_name)
}
\arguments{
\item{con}{Database connection}

\item{site_id}{Site ID}

\item{site_name}{Site name}
}
\description{
Insert Site
}
