% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_home.R
\name{mod_home_server}
\alias{mod_home_server}
\title{Home Page Module Server}
\usage{
mod_home_server(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Module server logic
}
\description{
Home Page Module Server
}
