% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sites.R
\name{mod_sites_ui}
\alias{mod_sites_ui}
\title{Sites Management Module UI - Placeholder}
\usage{
mod_sites_ui(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Shiny UI
}
\description{
Sites Management Module UI - Placeholder
}
