% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_utils.R
\name{validate_sample_id}
\alias{validate_sample_id}
\title{Validate Sample ID}
\usage{
validate_sample_id(sample_id)
}
\arguments{
\item{sample_id}{Sample ID to validate}
}
\value{
Logical TRUE if not empty
}
\description{
Validate Sample ID
}
