% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_utils.R
\name{validate_site_id}
\alias{validate_site_id}
\title{Validate Site ID}
\usage{
validate_site_id(site_id)
}
\arguments{
\item{site_id}{Site ID to validate}
}
\value{
Logical TRUE if valid format
}
\description{
Validate Site ID
}
