% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{add_site}
\alias{add_site}
\title{Add Site (Corrected version)}
\usage{
add_site(pool, site_name, description)
}
\arguments{
\item{pool}{Database pool connection}

\item{site_name}{Site name}

\item{description}{Site description}
}
\description{
Add Site (Corrected version)
}
