% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db.R
\name{get_next_site_id}
\alias{get_next_site_id}
\title{Get Next Sequential Site ID}
\usage{
get_next_site_id(con)
}
\arguments{
\item{con}{Database connection (or pool)}
}
\description{
Get Next Sequential Site ID
}
