% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sites.R
\name{mod_sites_server}
\alias{mod_sites_server}
\title{Sites Management Module Server - Placeholder}
\usage{
mod_sites_server(id)
}
\arguments{
\item{id}{Module ID}
}
\value{
Module server logic
}
\description{
Sites Management Module Server - Placeholder
}
